[move/#50] Detail 패키지 분리
